requirements.txt
numpy>=1.21.0
pygame>=2.1.0
librosa>=0.9.0
scipy>=1.7.0
opencv-python>=4.5.0
torch>=1.9.0
pillow>=8.3.0
tqdm>=4.62.0 